Inventory creation and slight player movement edit
"i" is the key used to open and close the inventory.
I created the inventory and functionality for spells to hold specific values related to how the function during gameplay.
Spells and inventory organization do not carry between runs of the game yet.
I fixed a small unnessisary repitition in the Alpha Movement Script, and added my inventory opening key to that script.
All of my changes can be found in the new inventory related scripts, prefabs, and canvas. The work and new GameObjects were all done in the "Test Area" scene, where I assume we will try to get a better grasp of the relationship between player movement/controls and UI.

Spells and inventory state still need to be saved to a file to give the player the same state when quitting and starting the game again.
"Test Area" needs some platforms to jump around.