loadout and NPC floating text
I made a loadout system that visually works, but the data is not yet being saved anywhere, and the inventory gets disabled when closed. Thus it is impossible to switch loadouts (even if we had keybinds for it), while playing rn.

added floating text that follows camera to test scene

upped the stim count for testing purposes, you are welcome XD