saving loadouts to file and box behind NPC text
I made functionality for saving which spells are in which loadout slots to a file. This information is ready to be hooked up with player inputs to allow the player to swap loadouts while the inventory is closed.

I also connected a variable from the Loadout script to this new LoadOutToFile script, so that the old one know what to display while the information in the file is incorrect (just meaning while the inventory is open).

also funny box behind NPC text